chore: core.py moved to core/dataset.py
